feat: UserName 객체 추가
