+++ README.md
@@ -36,4 +36,4 @@
 - [ ] given when then 스타일 테스트 작성
 - [ ] 방어적 복사 적용
 - [ ] 테스트 케이스 보강
-- [ ] 원시값 포장
+- [X] name String 원시값 포장
